Split ViewStateClassGenerator.java into separate classes
